requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Data acquisition
yfinance>=0.2.0
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Visualization
plotly>=5.17.0

# Statistics
statsmodels>=0.14.0


